Upgrade json, restforce, and webmock to latest versions.
The versions of each of these had fallen a reasonable way behind mainline, which could cause issues when trying to depend on this code in more up-to-date codebases.

There were 2 tests that I updated to make them more robust. The timeout one I made the queries explicit that were being mocked. There was another test that only worked up to the year 2020, so updating that to work until the year 3000.

The newer versions of Faraday now use a ConnectionFailed error to indicate a timeout when opening the connection (https://github.com/lostisland/faraday/issues/718), so updated accordingly.

Restforce 5 requires ruby 2.5, so updated that requirement as well. Also updated the tests to point to the latest version of the API at time of writing.